requirements.txt
altgraph==0.17
future==0.18.2
importlib-metadata==3.4.0
pefile==2019.4.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyOpenGL==3.1.5
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-snap7==0.11
pywin32-ctypes==0.2.0
typing-extensions==3.7.4.3
zipp==3.4.0
